Allows user to click on club to add it
Populates the createGroup page with clubs from cms, then allows the user
to simply click on a club to add it. For now it sends the user to the
/auth page if successful and does nothing if unsuccessful. It also,
doesn't stop clubs that are already in the database from showing up on
the createGroups page. Fix this, Future Tristan. Don't let your dreams
be dreams.